Remove redundant interface.
